pyproject.toml
[tool.poetry]
name = "drf_save_many_children"
version = "0.1.0"
description = ""
authors = ["Gubatenko Artem <gubatenko2011@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.2"
djangorestframework = "^3.12.1"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
